[feat/#48] Move expect file to package '/expect'
