+++ composeApp/src/jsMain/kotlin/com/survivai/survivai/expect/FontPreloader.web.kt
@@ -1,4 +1,4 @@
-package com.survivai.survivai
+package com.survivai.survivai.expect
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
